initial refactoring complete
